Log classnames of arguments in exception trace
